Auto generate raml endpoint
